Add supplemental features to the CSV export.
Also add some logic to retrieve the channel the spots have been
segmented from:

These features are taken from the spot features. Because we might
pair spots coming from different channels, we need to retrieve in
what channel the spots have been detected. Then we can retrieve the
feature in the right channel, for instance the mean of the channel 2
for the spots of the 2nd dataset.