Do not crash when the list of common spots is empty.
It is possible to have tracks that are matched with no common spots.
For instance, you can have a track 1 with spots at frames 1, 3 and 5,
and a track 2 with spots at frames 2 and 4. Using the average position
matching method will match these two tracks, but they will have no spots
in common. This happens with this method, and for tracks that have a lot
of gaps.

This commit simply avoid the crash and does not create a pair in these
cases.